Scroll na tela de detalhes, tela de favoritos, inicio coredata

+++ Movie Challenge/Movie Challenge/Services/dto/Movie.swift
@@ -60,5 +60,5 @@ struct Movie : Codable{
     var title: String?
     var video: Bool?
     var vote_average: Float?
-    var vote_count: Float?
+    var vote_count: Int?
 }
